refactor(trace): drop re-export shims and dead code
- chain_analysis: remove backward-compat wrappers delegating to order.py;
  import from order directly. Merge parse_source_for_try_except with its
  identical private twin.
- core: remove unused KEEP_AFTER_SUPPRESSION and EMPHASIS_MARKS constants
- typing: inline single-use RawChainItem alias
- trace/__init__: document why the ipython assignment exists (ty)
