+++ tracerite/trace/chain_analysis.py
@@ -2,7 +2,7 @@
 
 This module provides the pure try-except utilities used by the pipeline in
 ``tracerite.trace``.  The chronological-order construction itself lives in
-``trace.py`` so the whole pipeline is visible in one place.
+``order.py`` so the whole pipeline is visible in one place.
 """
 
 from __future__ import annotations
@@ -20,16 +20,13 @@ from .core import (
 )
 
 if TYPE_CHECKING:
-    from .typing import ChainLink, ExcChain, ExceptionInfo, FrameInfo, TryExceptBlock
+    from .typing import TryExceptBlock
 
 __all__ = [
     "parse_source_for_try_except",
     "parse_source_string_for_try_except",
     "find_try_block_for_except_line",
     "find_matching_try_for_inner_exception",
-    "analyze_exception_chain_links",
-    "enrich_chain_with_links",
-    "build_chronological_frames",
 ]
 
 
@@ -94,15 +91,6 @@ def parse_source_for_try_except(
     Returns:
         List of TryExceptBlock objects found in the source
     """
-    return _parse_source_for_try_except(filename, function_name, _cache=_cache)
-
-
-def _parse_source_for_try_except(
-    filename: str,
-    function_name: str | None = None,
-    *,
-    _cache: dict | None = None,
-) -> list[TryExceptBlock]:
     key = ("file", filename, function_name)
     if _cache is not None and key in _cache:
         return _cache[key]
@@ -193,67 +181,3 @@ def find_matching_try_for_inner_exception(
         ) and block_contains_in_try(block, inner_first_lineno):
             return block
     return None
-
-
-# The chronological-order builders live in tracerite.trace.order. The names
-# below are kept here so existing imports from ``tracerite.chain_analysis``
-# continue to work.
-
-
-def analyze_exception_chain_links(chain: ExcChain) -> list[ChainLink | None]:
-    """Analyze an exception chain to find try-except relationships."""
-    from .order import analyze_exception_chain_links
-
-    return analyze_exception_chain_links(chain)
-
-
-def enrich_chain_with_links(chain: ExcChain) -> ExcChain:
-    """Enrich exception chain with try-except link information."""
-    from .order import enrich_chain_with_links
-
-    return enrich_chain_with_links(chain)
-
-
-def build_chronological_frames(chain: ExcChain) -> list[FrameInfo]:
-    """Build a chronological list of frames showing the actual sequence of events."""
-    from .order import build_chronological_frames
-
-    return build_chronological_frames(chain)
-
-
-# Helpers that moved to tracerite.trace.order together with the pipeline code.
-# They are re-exported here for backward compatibility with existing tests.
-
-
-def get_frame_lineno(frame: FrameInfo) -> int | None:
-    from .order import get_frame_lineno
-
-    return get_frame_lineno(frame)
-
-
-def frame_in_except_handler(frame: FrameInfo) -> bool:
-    from .order import frame_in_except_handler
-
-    return frame_in_except_handler(frame)
-
-
-def find_chain_link(
-    inner_exc: ExceptionInfo, outer_exc: ExceptionInfo
-) -> ChainLink | None:
-    from .order import find_chain_link
-
-    return find_chain_link(inner_exc, outer_exc)
-
-
-def filter_hidden_frames(chronological: list[FrameInfo]) -> list[FrameInfo]:
-    from .order import filter_hidden_frames
-
-    return filter_hidden_frames(chronological)
-
-
-def apply_base_exception_suppression(
-    chronological: list[FrameInfo], chain: ExcChain
-) -> list[FrameInfo]:
-    from .order import apply_base_exception_suppression
-
-    return apply_base_exception_suppression(chronological, chain)
