+++ tracerite/trace/typing.py
@@ -20,5 +20,4 @@ ExceptionInfo: TypeAlias = dict[str, Any]
 ExcChain: TypeAlias = list[ExceptionInfo]
 Chain: TypeAlias = dict[str, Any]
 Fragment: TypeAlias = dict[str, Any]
-RawChainItem: TypeAlias = dict[str, Any]
-RawChain: TypeAlias = list[RawChainItem]
+RawChain: TypeAlias = list[dict[str, Any]]
